requirements.txt
flask 
numpy 
scipy 
pillow 
opencv-python
flask-cors
pydub